refactor: merged create
